Refactor static files serving routes registrations into a separate plugin.
