+++ components/log-viewer-webui/server/src/app.js
@@ -1,14 +1,11 @@
 import fastify from "fastify";
-import * as path from "node:path";
 import process from "node:process";
-import {fileURLToPath} from "node:url";
-
-import {fastifyStatic} from "@fastify/static";
 
 import settings from "../settings.json" with {type: "json"};
 import DbManager from "./DbManager.js";
 import exampleRoutes from "./routes/example.js";
 import queryRoutes from "./routes/query.js";
+import staticRoutes from "./routes/static.js";
 
 
 /**
@@ -26,30 +23,8 @@ const app = async ({
     sqlDbPass,
 }) => {
     const server = fastify(fastifyOptions);
-    const filename = fileURLToPath(import.meta.url);
-    const dirname = path.dirname(filename);
-    const parentDirname = path.resolve(dirname, "..");
 
     if ("test" !== process.env.NODE_ENV) {
-        let irFilesDir = settings.IrFilesDir;
-        if (false === path.isAbsolute(irFilesDir)) {
-            irFilesDir = path.resolve(parentDirname, irFilesDir);
-        }
-        await server.register(fastifyStatic, {
-            prefix: "/ir",
-            root: irFilesDir,
-        });
-
-        let logViewerDir = settings.LogViewerDir;
-        if (false === path.isAbsolute(logViewerDir)) {
-            logViewerDir = path.resolve(parentDirname, logViewerDir);
-        }
-        await server.register(fastifyStatic, {
-            prefix: "/log-viewer",
-            root: logViewerDir,
-            decorateReply: false,
-        });
-
         await server.register(DbManager, {
             mysqlConfig: {
                 database: settings.SqlDbName,
@@ -68,20 +43,7 @@ const app = async ({
         });
     }
 
-    if ("production" === process.env.NODE_ENV) {
-        // In the development environment, we expect the client to use a separate webserver that
-        // supports live reloading.
-        if (false === path.isAbsolute(settings.ClientDir)) {
-            throw new Error("`clientDir` must be an absolute path.");
-        }
-
-        await server.register(fastifyStatic, {
-            prefix: "/",
-            root: settings.ClientDir,
-            decorateReply: false,
-        });
-    }
-
+    await server.register(staticRoutes);
     await server.register(exampleRoutes);
     await server.register(queryRoutes);
